requirements.txt
sentence-transformers==3.0.1
numpy==2.0.2
pandas==2.2.2
python-docx==1.1.2
pypdf==4.2.0
requests==2.32.3
python-dotenv==1.0.0
kaggle==1.6.17
datasets==2.20.0
streamlit==1.39.0
